split components into tab sub-components
