+++ src/App.tsx
@@ -8,15 +8,15 @@ import { Profile } from './pages/Profile'
 import Build from './pages/Build'
 import View from './pages/View'
 import Tenants from './pages/Tenants'
-import CreateTenant from './pages/CreateTenant'
+import CreateTenant from './pages/create-tenant'
 import TenantStatus from './pages/TenantStatus'
 import TenantReadiness from './pages/TenantReadiness'
-import TenantDetails from './pages/TenantDetails'
+import TenantDetails from './pages/tenant-details'
 import AssignProjects from './pages/AssignProjects'
 import Projects from './pages/Projects'
 import CreateProject from './pages/CreateProject'
 import Administration from './pages/Administration'
-import ManageTenantMembers from './pages/ManageTenantMembers'
+import ManageTenantMembers from './pages/manage-tenant-members'
 import MyInvitations from './pages/MyInvitations'
 import { InvitationReview } from './pages/InvitationReview'
 import { Status } from './pages/Status'
