+++ src/pages/create-tenant/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CreateTenant'
