+++ src/pages/manage-tenant-members/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ManageTenantMembers'
