+++ src/pages/tenant-details/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TenantDetails'
